Finished product reviews

+++ src/main/java/com/revature/controllers/ProductReviewController.java
@@ -1,12 +1,15 @@
 package com.revature.controllers;
 
 import com.revature.annotations.Authorized;
+import com.revature.dtos.ProductReviewRequest;
 import com.revature.dtos.ProductReviewResponse;
 import com.revature.models.ProductReview;
+import com.revature.models.User;
 import com.revature.services.ProductReviewService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController
@@ -36,13 +39,20 @@ public class ProductReviewController {
 
     @Authorized
     @PutMapping
-    public ResponseEntity<ProductReviewResponse> upsert(@RequestBody ProductReview productReview) {
-        return ResponseEntity.ok(new ProductReviewResponse(productReviewService.save(productReview)));
+    public ResponseEntity<ProductReviewResponse> upsert(@RequestBody ProductReviewRequest productReview, HttpSession session) {
+        if(session.getAttribute("user") == null) {
+            return ResponseEntity.notFound().build();
+        }
+        User user = (User)session.getAttribute("user");
+        return ResponseEntity.ok(new ProductReviewResponse(productReviewService.save(productReview, user)));
     }
 
     @Authorized
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable("id") int id) {
+    public void deleteById(@PathVariable("id") int id, HttpSession session) {
+        if(session.getAttribute("user") == null) {
+            return;
+        }
         productReviewService.deleteById(id);
     }
 }
